requirements.txt
web3==6.0.0
python-dotenv==0.21.0
schedule==1.1.0
asyncio==3.4.3
